dnsdist: Fix compilation when dynamic blocks are disabled
