+++ pdns/dnsdistdist/dnsdist-lua.cc
@@ -38,12 +38,14 @@
 
 #include "dnsdist.hh"
 #include "dnsdist-backend.hh"
+#include "dnsdist-cache.hh"
 #include "dnsdist-carbon.hh"
 #include "dnsdist-concurrent-connections.hh"
 #include "dnsdist-configuration.hh"
 #include "dnsdist-console.hh"
 #include "dnsdist-crypto.hh"
 #include "dnsdist-dynblocks.hh"
+#include "dnsdist-dynbpf.hh"
 #include "dnsdist-discovery.hh"
 #include "dnsdist-ecs.hh"
 #include "dnsdist-frontend.hh"
